pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv-digest"
version = "0.1.0"
description = "Automated daily arXiv paper curation pipeline"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "PyMuPDF>=1.24",
    "google-genai>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

# ── Ruff Configuration ──

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
]
ignore = [
    "E501",  # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["arxiv_digest"]

# ── Pytest Configuration ──

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
